refactor: move engine related code to engine_db
Move the engine singleton and creat_engine function to the rest of the engine code
